# DMS_api_gateway/docs/swagger.yaml
basePath: /
definitions:
...
  entity.Putlist:
    properties:
      auto_id:
        type: integer
      bank_account_id:
        type: integer
      date_for:
        type: string
      date_with:
        type: string
      dispetcher_id:
        type: integer
      driver_id:
        type: integer
      id:
        type: integer
      mehanic_id:
        type: integer
      number:
        type: integer
      user_id:
        type: integer
    required:
    - auto_id
    - bank_account_id
    - date_for
    - date_with
    - dispetcher_id
    - driver_id
    - mehanic_id
    - number
    type: object
  entity.PutlistBody:
    properties:
      contragent_id:
        type: integer
      id:
        type: integer
      item:
        type: string
      number:
        type: integer
      putlist_number:
        type: integer
      time_for:
        type: string
      time_with:
        type: string
    required:
    - contragent_id
    - item
    - number
    - putlist_number
    - time_for
    - time_with
    type: object
...
  entity.UpdatePutlistBodyInput:
    properties:
      contragent_id:
        type: integer
      item:
        type: string
      number:
        type: integer
      time_for:
        type: string
      time_with:
        type: string
    type: object
  entity.UpdatePutlistHeaderInput:
    properties:
      auto_id:
        type: integer
      bank_account_id:
        type: integer
      date_for:
        type: string
      date_with:
        type: string
      dispetcher_id:
        type: integer
      driver_id:
        type: integer
      mehanic_id:
        type: integer
    type: object
...
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
...
  handler.getPutlistBodiesResponse:
    properties:
      putlist_bodies:
        items:
          $ref: '#/definitions/entity.PutlistBody'
        type: array
    type: object
  handler.getPutlistsByNumberResponse:
    properties:
      putlist:
        $ref: '#/definitions/entity.Putlist'
    type: object
  handler.getPutlistsResponse:
    properties:
      putlists:
        items:
          $ref: '#/definitions/entity.Putlist'
        type: array
    type: object
...
  handler.statusResponse:
    properties:
      status:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: API Server for Document Management Service
  title: DMS API
  version: "1.0"
paths:
  /{number}/putlist_bodies:
    get:
      consumes:
      - application/json
      description: gets data about all putlist bodies
      operationId: get-putlist-bodies
      parameters:
      - description: putlist number
        in: path
        name: number
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: putlist bodies info
          schema:
            $ref: '#/definitions/handler.getPutlistBodiesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Putlist Bodies
      tags:
      - putlist-bodies
    post:
      consumes:
      - application/json
      description: creates and adds data about the putlist body
      operationId: create-putlist-body
      parameters:
      - description: putlist number
        in: path
        name: number
        required: true
        type: integer
      - description: putlist body info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/entity.PutlistBody'
      produces:
      - application/json
      responses:
        "200":
          description: putlist body ID
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Putlist Body
      tags:
      - putlist-bodies
  /{number}/putlist_bodies/{putlist_body_id}:
    delete:
      consumes:
      - application/json
      description: deletes data about putlist body
      operationId: delete-putlist-body
      parameters:
      - description: putlist number
        in: path
        name: number
        required: true
        type: integer
      - description: putlist body id
        in: path
        name: putlist_body_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: deleted
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Putlist Body
      tags:
      - putlist-bodies
    put:
      consumes:
      - application/json
      description: updates data about the putlist body
      operationId: update-putlist-body
      parameters:
      - description: putlist number
        in: path
        name: number
        required: true
        type: integer
      - description: putlist body id
        in: path
        name: putlist_body_id
        required: true
        type: integer
      - description: update putlist body info
        in: body
        name: updateData
        required: true
        schema:
          $ref: '#/definitions/entity.UpdatePutlistBodyInput'
      produces:
      - application/json
      responses:
        "200":
          description: updated putlist body
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Putlist Body
      tags:
      - putlist-bodies
...
  /putlists:
    get:
      consumes:
      - application/json
      description: gets data about all putlist headers
      operationId: get-putlists
      produces:
      - application/json
      responses:
        "200":
          description: putlist headers info
          schema:
            $ref: '#/definitions/handler.getPutlistsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Putlists
      tags:
      - putlists
    post:
      consumes:
      - application/json
      description: creates and adds data about the putlist
      operationId: create-putlist
      parameters:
      - description: putlist info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/entity.Putlist'
      produces:
      - application/json
      responses:
        "200":
          description: putlist ID
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Putlist
      tags:
      - putlists
  /putlists/{number}:
    delete:
      consumes:
      - application/json
      description: deletes data about putlist
      operationId: delete-putlist
      parameters:
      - description: number
        in: path
        name: number
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: deleted
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Putlist
      tags:
      - putlists
    get:
      consumes:
      - application/json
      description: gets data about putlist
      operationId: get-putlist-by-number
      parameters:
      - description: number
        in: path
        name: number
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: putlist header info
          schema:
            $ref: '#/definitions/handler.getPutlistsByNumberResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Putlist By Number
      tags:
      - putlists
    put:
      consumes:
      - application/json
      description: updates data about the putlist header
      operationId: update-putlist
      parameters:
      - description: number
        in: path
        name: number
        required: true
        type: integer
      - description: update putlist info
        in: body
        name: updateData
        required: true
        schema:
          $ref: '#/definitions/entity.UpdatePutlistHeaderInput'
      produces:
      - application/json
      responses:
        "200":
          description: updated putlist
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Putlist
      tags:
      - putlists
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"